+++ packages/odata-service-writer/test/unit/index.test.ts
@@ -1,5 +1,5 @@
-import type { OdataService } from '../../src';
-import { generate, OdataVersion, ServiceType } from '../../src';
+import type { EdmxAnnotationsInfo, OdataService } from '../../src';
+import { generate, remove, OdataVersion, ServiceType } from '../../src';
 import { join } from 'path';
 import type { Editor } from 'mem-fs-editor';
 import { create } from 'mem-fs-editor';
@@ -160,10 +160,53 @@ describe('generate', () => {
             expect(manifest['sap.app'].dataSources.mainService.uri).toBe(config.path);
             expect(manifest['sap.app'].dataSources[config.annotations.technicalName]).toBeDefined();
             // verify local copy of metadata
-            expect(fs.read(join(testDir, 'webapp', 'localService', 'metadata.xml'))).toBe(config.metadata);
-            expect(fs.read(join(testDir, 'webapp', 'localService', `${config.annotations.technicalName}.xml`))).toBe(
-                config.annotations.xml
+            expect(fs.read(join(testDir, 'webapp', 'localService', 'mainService', 'metadata.xml'))).toBe(
+                config.metadata
             );
+            expect(
+                fs.read(
+                    join(testDir, 'webapp', 'localService', 'mainService', `${config.annotations.technicalName}.xml`)
+                )
+            ).toBe(config.annotations.xml);
+        });
+
+        it('Valid OData V2 service with multiple annotations', async () => {
+            const config = {
+                ...commonConfig,
+                version: OdataVersion.v2,
+                annotations: [
+                    {
+                        technicalName: 'TEST_ME',
+                        xml: '<HELLO WORLD />'
+                    },
+                    {
+                        technicalName: 'TEST_ME_TWO',
+                        xml: '<HELLO WORLD TWO />'
+                    }
+                ]
+            };
+            await generate(testDir, config as OdataService, fs);
+
+            // verify updated manifest.json
+            const manifest = fs.readJSON(join(testDir, 'webapp', 'manifest.json')) as any;
+            expect(manifest['sap.app'].dataSources.mainService.uri).toBe(config.path);
+            expect(manifest['sap.app'].dataSources[config.annotations[0].technicalName]).toBeDefined();
+            expect(manifest['sap.app'].dataSources[config.annotations[1].technicalName]).toBeDefined();
+            // verify local copy of metadata
+            // mainService should be used in case there is no name defined for service
+            expect(fs.read(join(testDir, 'webapp', 'localService', 'mainService', 'metadata.xml'))).toBe(
+                config.metadata
+            );
+            expect(
+                fs.read(
+                    join(testDir, 'webapp', 'localService', 'mainService', `${config.annotations[0].technicalName}.xml`)
+                )
+            ).toBe(config.annotations[0].xml);
+            expect(
+                fs.read(
+                    join(testDir, 'webapp', 'localService', 'mainService', `${config.annotations[1].technicalName}.xml`)
+                )
+            ).toBe(config.annotations[1].xml);
         });
 
         it('Valid OData V4 service', async () => {
@@ -178,7 +221,12 @@ describe('generate', () => {
             const manifest = fs.readJSON(join(testDir, 'webapp', 'manifest.json')) as any;
             expect(manifest['sap.app'].dataSources[config.name].uri).toBe(config.path);
             // verify local copy of metadata
-            expect(fs.read(join(testDir, 'webapp', 'localService', 'metadata.xml'))).toBe(config.metadata);
+            // first service is always mainService, so we make sure data for it is generated in correct location
+            expect(fs.exists(join(testDir, 'webapp', 'localService', 'myService', 'metadata.xml'))).toBe(false);
+            expect(fs.exists(join(testDir, 'webapp', 'localService', 'mainService', 'metadata.xml'))).toBe(true);
+            expect(fs.read(join(testDir, 'webapp', 'localService', 'mainService', 'metadata.xml'))).toBe(
+                config.metadata
+            );
             // verify that no destination is added to the ui5.yaml
             expect(fs.read(join(testDir, 'ui5.yaml'))).not.toContain('destination: ');
             // verify that client is set
@@ -276,9 +324,9 @@ describe('generate', () => {
                 path: '/V2/Northwind/Northwind.svc',
                 version: OdataVersion.v2
             } as OdataService;
-
+            // No services are defined - mainService used for service name and '' for service model
             let configCopy = cloneDeep(config);
-            enhanceData(configCopy);
+            await enhanceData('', configCopy, fs);
             expect(configCopy).toMatchInlineSnapshot(`
                 Object {
                   "model": "",
@@ -294,8 +342,13 @@ describe('generate', () => {
                 }
             `);
 
+            // Services already defined - actual service name and service model are used
             configCopy = cloneDeep(Object.assign({}, config, { model: 'modelName', name: 'datasourceName' }));
-            enhanceData(configCopy);
+            fs.writeJSON(join('webapp', 'manifest.json'), {
+                'sap.app': { dataSources: { existingService: { type: 'OData' } } },
+                'sap.ui5': { models: { existingModel: { dataSource: 'existingService' } } }
+            });
+            await enhanceData('', configCopy, fs);
             expect(configCopy).toMatchInlineSnapshot(`
                 Object {
                   "model": "modelName",
@@ -311,9 +364,10 @@ describe('generate', () => {
                 }
             `);
 
+            fs.delete(join('webapp', 'manifest.json'));
             // Undefined path does not throw but sets valid path
             configCopy = cloneDeep(Object.assign({}, config, { path: undefined }));
-            enhanceData(configCopy);
+            await enhanceData('', configCopy, fs);
             expect(configCopy).toMatchInlineSnapshot(`
                 Object {
                   "model": "",
@@ -328,6 +382,199 @@ describe('generate', () => {
                   "version": "2",
                 }
             `);
+
+            // Service and annotation names are the same
+            fs.writeJSON(join('webapp', 'manifest.json'), {
+                'sap.app': { dataSources: { exisitingSerivce: { type: 'OData' } } }
+            });
+            configCopy = cloneDeep(Object.assign({}, config, { name: 'aname', annotations: { name: 'aname' } }));
+            await enhanceData('', configCopy, fs);
+            expect(configCopy).toMatchInlineSnapshot(`
+                Object {
+                  "annotations": Object {
+                    "name": "aname_Annotation",
+                  },
+                  "model": "",
+                  "name": "aname",
+                  "path": "/V2/Northwind/Northwind.svc/",
+                  "previewSettings": Object {
+                    "path": "/V2",
+                    "url": "https://services.odata.org",
+                  },
+                  "type": "edmx",
+                  "url": "https://services.odata.org",
+                  "version": "2",
+                }
+            `);
         });
+    });
+});
+
+describe('remove', () => {
+    let fs: Editor;
+    beforeEach(async () => {
+        const ui5Yaml = (await UI5Config.newInstance(''))
+            .addFioriToolsProxydMiddleware({ ui5: {}, backend: [{ path: '/sap', url: 'https://localhost' }] })
+            .toString();
+        const ui5LocalYaml = (await UI5Config.newInstance(''))
+            .addFioriToolsProxydMiddleware({ ui5: {}, backend: [{ path: '/sap', url: 'https://localhost' }] })
+            .addMockServerMiddleware(
+                [{ serviceName: 'mainService', servicePath: '/sap' }],
+                [
+                    {
+                        urlPath: `/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/`
+                    }
+                ]
+            )
+            .toString();
+        const ui5MockYaml = (await UI5Config.newInstance(''))
+            .addFioriToolsProxydMiddleware({ ui5: {}, backend: [{ path: '/sap', url: 'https://localhost' }] })
+            .addMockServerMiddleware(
+                [{ serviceName: 'mainService', servicePath: '/sap' }],
+                [
+                    {
+                        urlPath: `/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/`
+                    }
+                ]
+            )
+            .toString();
+        // generate required files
+        fs = create(createStorage());
+        fs.write(join(testDir, 'ui5.yaml'), ui5Yaml);
+        fs.write(join(testDir, 'ui5-local.yaml'), ui5LocalYaml);
+        fs.write(join(testDir, 'ui5-mock.yaml'), ui5MockYaml);
+        fs.writeJSON(join(testDir, 'package.json'), { ui5: { dependencies: [] } });
+        fs.write(
+            join(testDir, 'webapp', 'manifest.json'),
+            JSON.stringify({
+                'sap.app': {
+                    id: 'testappid',
+                    dataSources: {
+                        mainService: {
+                            uri: '/sap',
+                            type: 'OData',
+                            settings: {
+                                annotations: ['SEPMRA_PROD_MAN', 'annotation'],
+                                localUri: 'annotations/annotation.xml'
+                            }
+                        },
+                        SEPMRA_PROD_MAN: {
+                            uri: `/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/`,
+                            type: 'ODataAnnotation',
+                            settings: {
+                                localUri: 'localService/SEPMRA_PROD_MAN.xml'
+                            }
+                        },
+                        annotation: {
+                            type: 'ODataAnnotation',
+                            uri: 'annotations/annotation.xml',
+                            settings: {
+                                localUri: 'annotations/annotation.xml'
+                            }
+                        }
+                    }
+                },
+                'sap.ui5': {
+                    models: {
+                        '': {
+                            dataSource: 'mainService'
+                        }
+                    }
+                }
+            })
+        );
+    });
+    it('Try to remove an unexisting service', async () => {
+        await remove(
+            testDir,
+            {
+                name: 'dummyService',
+                url: 'https://dummyUrl',
+                path: '/dummyPath',
+                type: ServiceType.EDMX,
+                annotations: [{ technicalName: 'dummy-technical-name' }] as EdmxAnnotationsInfo[]
+            },
+            fs
+        );
+        // verify updated manifest.json
+        const manifest = fs.readJSON(join(testDir, 'webapp', 'manifest.json')) as any;
+        expect(manifest['sap.app'].dataSources).toStrictEqual({
+            mainService: {
+                uri: '/sap',
+                type: 'OData',
+                settings: {
+                    annotations: ['SEPMRA_PROD_MAN', 'annotation'],
+                    localUri: 'annotations/annotation.xml'
+                }
+            },
+            SEPMRA_PROD_MAN: {
+                uri: `/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/`,
+                type: 'ODataAnnotation',
+                settings: {
+                    localUri: 'localService/SEPMRA_PROD_MAN.xml'
+                }
+            },
+            annotation: {
+                type: 'ODataAnnotation',
+                uri: 'annotations/annotation.xml',
+                settings: {
+                    localUri: 'annotations/annotation.xml'
+                }
+            }
+        });
+        expect(manifest['sap.ui5'].models).toStrictEqual({
+            '': {
+                dataSource: 'mainService'
+            }
+        });
+        // verify ui5.yaml, ui5-local.yaml, ui5-mock.yaml
+        expect(fs.read(join(testDir, 'ui5.yaml'))).toContain(
+            'backend:\n          - path: /sap\n            url: https://localhost\n'
+        );
+        expect(fs.read(join(testDir, 'ui5-local.yaml'))).toContain(
+            'backend:\n          - path: /sap\n            url: https://localhost\n'
+        );
+        expect(fs.read(join(testDir, 'ui5-mock.yaml'))).toContain('services:\n          - urlPath: /sap\n');
+        expect(fs.read(join(testDir, 'ui5-mock.yaml'))).toContain(
+            `annotations:\n          - urlPath: /sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/\n`
+        );
+    });
+    it('Remove an existing service', async () => {
+        await remove(
+            testDir,
+            {
+                name: 'mainService',
+                url: 'https://localhost',
+                path: '/sap',
+                type: ServiceType.EDMX,
+                annotations: [
+                    { technicalName: 'SEPMRA_PROD_MAN', xml: '<?xml version="1.0" encoding="utf-8"?></edmx:Edmx>' }
+                ] as EdmxAnnotationsInfo[]
+            },
+            fs
+        );
+        // verify updated manifest.json
+        const manifest = fs.readJSON(join(testDir, 'webapp', 'manifest.json')) as any;
+        expect(manifest['sap.app'].dataSources).toStrictEqual({
+            annotation: {
+                type: 'ODataAnnotation',
+                uri: 'annotations/annotation.xml',
+                settings: {
+                    localUri: 'annotations/annotation.xml'
+                }
+            }
+        });
+        expect(manifest['sap.ui5'].models).toStrictEqual({});
+        // verify ui5.yaml, ui5-local.yaml, ui5-mock.yaml
+        expect(fs.read(join(testDir, 'ui5.yaml'))).not.toContain('- path: /sap\n            url: http://localhost\n');
+        expect(fs.read(join(testDir, 'ui5-local.yaml'))).not.toContain(
+            '- path: /sap\n            url: http://localhost\n'
+        );
+        expect(fs.read(join(testDir, 'ui5-mock.yaml'))).not.toContain(
+            'services:\n          - urlPath: /sap/odata/testme\n '
+        );
+        expect(fs.read(join(testDir, 'ui5-mock.yaml'))).not.toContain(
+            `annotations:\n          - urlPath: /sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/\n`
+        );
     });
 });